Add Voter

+++ src/Controller/Web/DeleteUser/v1/Manager.php
@@ -4,10 +4,16 @@ namespace App\Controller\Web\DeleteUser\v1;
 
 use App\Domain\Entity\User;
 use App\Domain\Service\UserService;
+use App\Application\Security\Voter\UserVoter;
+use App\Controller\Exception\AccessDeniedException;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
 class Manager
 {
-    public function __construct(private readonly UserService $userService)
+    public function __construct(
+        private readonly UserService $userService,
+        private readonly AuthorizationCheckerInterface $authorizationChecker
+    )
     {
     }
 
@@ -18,6 +24,10 @@ class Manager
 
     public function deleteUser(User $user): void
     {
+        if (!$this->authorizationChecker->isGranted(UserVoter::DELETE, $user)) {
+            throw new AccessDeniedException();
+        }
+
         $this->userService->remove($user);
     }
 }
